Convert Prisma migrations from SQLite to PostgreSQL syntax
- Updated DATETIME to TIMESTAMP(3)
- Updated REAL to DOUBLE PRECISION
- Changed PRIMARY KEY syntax to PostgreSQL format
- Removed SQLite PRAGMA statements
- Added proper CONSTRAINT declarations for foreign keys

+++ prisma/migrations/20251209204736_add_order_tracking/migration.sql
@@ -1,3 +1,3 @@
 -- AlterTable
 ALTER TABLE "Participant" ADD COLUMN "ordersCount" INTEGER;
-ALTER TABLE "Participant" ADD COLUMN "totalSpent" REAL;
+ALTER TABLE "Participant" ADD COLUMN "totalSpent" DOUBLE PRECISION;
